Fix type duplication by moving MCPServerConfig to shared types
Creates src/types/mcp.ts as the single source of truth for MCPServerConfig,
addressing the medium-priority PR review comment about type duplication.

- types/mcp.ts: New file defining MCPServerConfig
- types/engine.ts: Now imports from types/mcp.ts
- services/mcp/types.ts: Re-exports from types/mcp.ts to maintain API

This avoids circular imports while eliminating duplicate type definitions.
